src-py/requirements.txt
fastapi
sqlmodel
pydantic
uvicorn[standard]
python-multipart
pytest
pytest-asyncio
httpx
alembic
python-dotenv
chromadb
